refactor: 코드 및 파일 구조 정리
